doi ten package hieuapp -> android, xoa comment hieuapp
