+++ RivChat/app/src/main/java/com/android/rivchat/model/Consersation.java
@@ -1,0 +1,16 @@
+package com.android.rivchat.model;
+
+import java.util.ArrayList;
+
+
+
+public class Consersation {
+    private ArrayList<Message> listMessageData;
+    public Consersation(){
+        listMessageData = new ArrayList<>();
+    }
+
+    public ArrayList<Message> getListMessageData() {
+        return listMessageData;
+    }
+}
