+++ RivChat/app/src/main/java/com/android/rivchat/model/Friend.java
@@ -1,0 +1,8 @@
+package com.android.rivchat.model;
+
+
+
+public class Friend extends User{
+    public String id;
+    public String idRoom;
+}
