+++ RivChat/app/src/main/java/com/android/rivchat/model/Group.java
@@ -1,0 +1,12 @@
+package com.android.rivchat.model;
+
+
+
+public class Group extends Room{
+    public String id;
+    public ListFriend listFriend;
+
+    public Group(){
+        listFriend = new ListFriend();
+    }
+}
