+++ RivChat/app/src/main/java/com/android/rivchat/model/ListFriend.java
@@ -1,0 +1,29 @@
+package com.android.rivchat.model;
+
+import java.util.ArrayList;
+
+
+public class ListFriend {
+    private ArrayList<Friend> listFriend;
+
+    public ArrayList<Friend> getListFriend() {
+        return listFriend;
+    }
+
+    public ListFriend(){
+        listFriend = new ArrayList<>();
+    }
+
+    public String getAvataById(String id){
+        for(Friend friend: listFriend){
+            if(id.equals(friend.id)){
+                return friend.avata;
+            }
+        }
+        return "";
+    }
+
+    public void setListFriend(ArrayList<Friend> listFriend) {
+        this.listFriend = listFriend;
+    }
+}
