+++ RivChat/app/src/main/java/com/android/rivchat/model/Message.java
@@ -1,0 +1,10 @@
+package com.android.rivchat.model;
+
+
+
+public class Message{
+    public String idSender;
+    public String idReceiver;
+    public String text;
+    public long timestamp;
+}
