+++ RivChat/app/src/main/java/com/android/rivchat/model/Room.java
@@ -1,0 +1,17 @@
+package com.android.rivchat.model;
+
+import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.Map;
+
+
+
+public class Room {
+    public ArrayList<String> member;
+    public Map<String, String> groupInfo;
+
+    public Room(){
+        member = new ArrayList<>();
+        groupInfo = new HashMap<String, String>();
+    }
+}
