+++ RivChat/app/src/main/java/com/android/rivchat/model/Status.java
@@ -1,0 +1,13 @@
+package com.android.rivchat.model;
+
+
+
+public class Status{
+    public boolean isOnline;
+    public long timestamp;
+
+    public Status(){
+        isOnline = false;
+        timestamp = 0;
+    }
+}
