+++ RivChat/app/src/main/java/com/android/rivchat/model/User.java
@@ -1,0 +1,23 @@
+package com.android.rivchat.model;
+
+
+
+public class User {
+    public String name;
+    public String email;
+    public String avata;
+    public Status status;
+    public Message message;
+
+
+    public User(){
+        status = new Status();
+        message = new Message();
+        status.isOnline = false;
+        status.timestamp = 0;
+        message.idReceiver = "0";
+        message.idSender = "0";
+        message.text = "";
+        message.timestamp = 0;
+    }
+}
