--- RivChat/app/src/main/java/com/hieuapp/rivchat/data/SharedPreferenceHelper.java
@@ -1,59 +1,0 @@
-package com.hieuapp.rivchat.data;
-
-import android.content.Context;
-import android.content.SharedPreferences;
-
-import com.hieuapp.rivchat.model.User;
-
-/**
- * Created by nguyenbinh on 09/12/2016.
- */
-
-public class SharedPreferenceHelper {
-    private static SharedPreferenceHelper instance = null;
-    private static SharedPreferences preferences;
-    private static SharedPreferences.Editor editor;
-    private static String SHARE_USER_INFO = "userinfo";
-    private static String SHARE_KEY_NAME = "name";
-    private static String SHARE_KEY_EMAIL = "email";
-    private static String SHARE_KEY_AVATA = "avata";
-    private static String SHARE_KEY_UID = "uid";
-
-
-    private SharedPreferenceHelper() {}
-
-    public static SharedPreferenceHelper getInstance(Context context) {
-        if (instance == null) {
-            instance = new SharedPreferenceHelper();
-            preferences = context.getSharedPreferences(SHARE_USER_INFO, Context.MODE_PRIVATE);
-            editor = preferences.edit();
-        }
-        return instance;
-    }
-
-    public void saveUserInfo(User user) {
-        editor.putString(SHARE_KEY_NAME, user.name);
-        editor.putString(SHARE_KEY_EMAIL, user.email);
-        editor.putString(SHARE_KEY_AVATA, user.avata);
-        editor.putString(SHARE_KEY_UID, StaticConfig.UID);
-        editor.apply();
-    }
-
-    public User getUserInfo(){
-        String userName = preferences.getString(SHARE_KEY_NAME, "");
-        String email = preferences.getString(SHARE_KEY_EMAIL, "");
-        String avatar = preferences.getString(SHARE_KEY_AVATA, "default");
-
-        User user = new User();
-        user.name = userName;
-        user.email = email;
-        user.avata = avatar;
-
-        return user;
-    }
-
-    public String getUID(){
-        return preferences.getString(SHARE_KEY_UID, "");
-    }
-
-}
