--- RivChat/app/src/main/java/com/hieuapp/rivchat/model/Configuration.java
@@ -1,29 +1,0 @@
-package com.hieuapp.rivchat.model;
-
-/**
- * Created by hieuttc on 05/12/2016.
- */
-
-public class Configuration {
-    private String label;
-    private String  value;
-    private int icon;
-
-    public Configuration(String label, String value, int icon){
-        this.label = label;
-        this.value = value;
-        this.icon = icon;
-    }
-
-    public String getLabel(){
-        return this.label;
-    }
-
-    public String getValue(){
-        return this.value;
-    }
-
-    public int getIcon(){
-        return this.icon;
-    }
-}
