--- RivChat/app/src/main/java/com/hieuapp/rivchat/model/Consersation.java
@@ -1,18 +1,0 @@
-package com.hieuapp.rivchat.model;
-
-import java.util.ArrayList;
-
-/**
- * Created by nguyenbinh on 10/12/2016.
- */
-
-public class Consersation {
-    private ArrayList<Message> listMessageData;
-    public Consersation(){
-        listMessageData = new ArrayList<>();
-    }
-
-    public ArrayList<Message> getListMessageData() {
-        return listMessageData;
-    }
-}
