--- RivChat/app/src/main/java/com/hieuapp/rivchat/model/Friend.java
@@ -1,10 +1,0 @@
-package com.hieuapp.rivchat.model;
-
-/**
- * Created by nguyenbinh on 10/12/2016.
- */
-
-public class Friend extends User{
-    public String id;
-    public String idRoom;
-}
