--- RivChat/app/src/main/java/com/hieuapp/rivchat/model/Message.java
@@ -1,12 +1,0 @@
-package com.hieuapp.rivchat.model;
-
-/**
- * Created by nguyenbinh on 08/12/2016.
- */
-
-public class Message{
-    public String idSender;
-    public String idReceiver;
-    public String text;
-    public long timestamp;
-}
