--- RivChat/app/src/main/java/com/hieuapp/rivchat/model/Status.java
@@ -1,15 +1,0 @@
-package com.hieuapp.rivchat.model;
-
-/**
- * Created by nguyenbinh on 08/12/2016.
- */
-
-public class Status{
-    public boolean isOnline;
-    public long timestamp;
-
-    public Status(){
-        isOnline = false;
-        timestamp = 0;
-    }
-}
