--- RivChat/app/src/main/java/com/hieuapp/rivchat/receiver/NetworkReceiver.java
@@ -1,16 +1,0 @@
-package com.hieuapp.rivchat.receiver;
-
-import android.content.BroadcastReceiver;
-import android.content.Context;
-import android.content.Intent;
-
-/**
- * Created by hieuttc on 05/12/2016.
- */
-
-public class NetworkReceiver extends BroadcastReceiver {
-    @Override
-    public void onReceive(Context context, Intent intent) {
-
-    }
-}
