--- RivChat/app/src/main/java/com/hieuapp/rivchat/service/GroupChatService.java
@@ -1,8 +1,0 @@
-package com.hieuapp.rivchat.service;
-
-/**
- * Created by nguyenbinh on 12/12/2016.
- */
-
-public class GroupChatService {
-}
