--- RivChat/app/src/main/java/com/hieuapp/rivchat/ui/ListBaseActivity.java
@@ -1,8 +1,0 @@
-package com.hieuapp.rivchat.ui;
-
-/**
- * Created by hieuttc on 05/12/2016.
- */
-
-public class ListBaseActivity {
-}
